backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.6.1
dnspython==2.4.2
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
bcrypt==4.1.2
requests==2.31.0
anthropic==0.8.1
openai==1.6.1